Move paging code from Core & Mgmt.Models into Mgmt.Paging
