+++ src/Auth0.ManagementApi/Clients/ClientsClient.cs
@@ -1,7 +1,6 @@
-using Auth0.Core.Collections;
 using Auth0.Core.Http;
 using Auth0.ManagementApi.Models;
-using Auth0.ManagementApi.Serialization;
+using Auth0.ManagementApi.Paging;
 using System;
 using System.Collections.Generic;
 using System.Linq;
