+++ tests/Auth0.ManagementApi.IntegrationTests/ClientTests.cs
@@ -2,6 +2,7 @@
 using System.Threading.Tasks;
 using Auth0.Core.Exceptions;
 using Auth0.ManagementApi.Models;
+using Auth0.ManagementApi.Paging;
 using Auth0.Tests.Shared;
 using FluentAssertions;
 using Xunit;
