+++ tests/Auth0.ManagementApi.IntegrationTests/RolesTests.cs
@@ -7,6 +7,7 @@ using Auth0.ManagementApi.Models;
 using FluentAssertions;
 using Xunit;
 using Auth0.Tests.Shared;
+using Auth0.ManagementApi.Paging;
 
 namespace Auth0.ManagementApi.IntegrationTests
 {
